python/requirements.txt
requests
infi.systray